pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "assemblyai>=0.48.2",
    "chromadb>=1.3.4",
    "cryptography>=46.0.3",
    "diffusers>=0.36.0",
    "docling>=2.58.0",
    "docx>=0.2.4",
    "docx2txt>=0.9",
    "easyocr>=1.7.2",
    "fastapi[standard]>=0.120.0",
    "faster-whisper>=1.2.1",
    "ffmpeg>=1.4",
    "ffmpeg-python>=0.2.0",
    "langchain>=1.0.5",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-experimental>=0.4.0",
    "langchain-google-genai>=3.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "langchain-text-splitters>=1.0.0",
    "mysql>=0.0.3",
    "openai-whisper>=20250625",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pillow>=10.4.0",
    "pipecat-ai>=0.0.97",
    "piper-tts>=1.3.0",
    "pyaudio>=0.2.14",
    "pydantic>=2.12.3",
    "pymupdf>=1.26.6",
    "pymupdf4llm>=0.1.8",
    "pymysql>=1.1.2",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
    "pytube>=15.0.0",
    "rarfile>=4.2",
    "scalar-fastapi>=1.4.4",
    "scipy>=1.16.3",
    "sentence-transformers>=5.1.2",
    "speechrecognition[pocketsphinx,vosk]>=3.14.4",
    "sqlalchemy>=2.0.44",
    "tiktoken>=0.12.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
    "vosk>=0.3.45",
    "xlrd>=2.0.2",
    "youtube-transcript-api>=1.2.3",
    "yt-dlp>=2024.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.14.4",
]

